Refactoring of distributions module, work in progress
